+++ src/zenml/zen_stores/schemas/artifact_schemas.py
@@ -13,9 +13,8 @@
 #  permissions and limitations under the License.
 """SQLModel implementation of artifact table."""
 
-import json
 from datetime import datetime
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import TYPE_CHECKING, Any, List, Optional
 from uuid import UUID
 
 from pydantic import ValidationError
@@ -30,7 +29,6 @@ from zenml.enums import (
     MetadataResourceTypes,
     TaggableResourceTypes,
 )
-from zenml.metadata.metadata_types import MetadataType
 from zenml.models import (
     ArtifactResponse,
     ArtifactResponseBody,
@@ -41,10 +39,12 @@ from zenml.models import (
     ArtifactVersionResponseBody,
     ArtifactVersionResponseMetadata,
     ArtifactVersionUpdate,
-    RunMetadataEntry,
 )
 from zenml.models.v2.core.artifact import ArtifactRequest
-from zenml.zen_stores.schemas.base_schemas import BaseSchema, NamedSchema
+from zenml.zen_stores.schemas.base_schemas import (
+    BaseSchema,
+    NamedSchema,
+)
 from zenml.zen_stores.schemas.component_schemas import StackComponentSchema
 from zenml.zen_stores.schemas.schema_utils import build_foreign_key_field
 from zenml.zen_stores.schemas.step_run_schemas import (
@@ -52,6 +52,7 @@ from zenml.zen_stores.schemas.step_run_schemas import (
     StepRunOutputArtifactSchema,
 )
 from zenml.zen_stores.schemas.user_schemas import UserSchema
+from zenml.zen_stores.schemas.utils import RunMetadataInterface
 from zenml.zen_stores.schemas.workspace_schemas import WorkspaceSchema
 
 if TYPE_CHECKING:
@@ -175,7 +176,7 @@ class ArtifactSchema(NamedSchema, table=True):
         return self
 
 
-class ArtifactVersionSchema(BaseSchema, table=True):
+class ArtifactVersionSchema(BaseSchema, RunMetadataInterface, table=True):
     """SQL Model for artifact versions."""
 
     __tablename__ = "artifact_version"
@@ -309,41 +310,6 @@ class ArtifactVersionSchema(BaseSchema, table=True):
             data_type=artifact_version_request.data_type.model_dump_json(),
             save_type=artifact_version_request.save_type.value,
         )
-
-    def fetch_metadata_collection(self) -> Dict[str, List[RunMetadataEntry]]:
-        """Fetches all the metadata entries related to the artifact version.
-
-        Returns:
-            a dictionary, where the key is the key of the metadata entry
-                and the values represent the list of entries with this key.
-        """
-        metadata_collection: Dict[str, List[RunMetadataEntry]] = {}
-
-        # Fetch the metadata related to this step
-        for rm in self.run_metadata_resources:
-            if rm.run_metadata.key not in metadata_collection:
-                metadata_collection[rm.run_metadata.key] = []
-            metadata_collection[rm.run_metadata.key].append(
-                RunMetadataEntry(
-                    value=json.loads(rm.run_metadata.value),
-                    created=rm.run_metadata.created,
-                )
-            )
-
-        return metadata_collection
-
-    def fetch_metadata(self) -> Dict[str, MetadataType]:
-        """Fetches the latest metadata entry related to the artifact version.
-
-        Returns:
-            a dictionary, where the key is the key of the metadata entry
-                and the values represent the latest entry with this key.
-        """
-        metadata_collection = self.fetch_metadata_collection()
-        return {
-            k: sorted(v, key=lambda x: x.created, reverse=True)[0].value
-            for k, v in metadata_collection.items()
-        }
 
     def to_model(
         self,
